UI Improved ListView of Request finished Todo: get requests from firebase Onclick Actions (Firebase delete, add user to following follower list) ....
