+++ app/src/main/java/com/example/goalog/UserPageActivity.java
@@ -180,7 +180,7 @@ public class UserPageActivity extends AppCompatActivity {
                         overridePendingTransition(0, 0);
                         return true;
                     case R.id.navigation_notifications:
-                        startActivity(new Intent(getApplicationContext(), NotificationActivity.class));
+                        startActivity(new Intent(getApplicationContext(), RequestActivity.class));
                         overridePendingTransition(0, 0);
                         return true;
                     case R.id.navigation_my_page:
