+++ app/src/main/res/drawable/ic_close_round.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="122.881dp"
+    android:height="122.88dp"
+    android:viewportWidth="122.881"
+    android:viewportHeight="122.88">
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M61.44,0c33.933,0 61.441,27.507 61.441,61.439c0,33.933 -27.508,61.44 -61.441,61.44C27.508,122.88 0,95.372 0,61.439C0,27.507 27.508,0 61.44,0L61.44,0zM81.719,36.226c1.363,-1.363 3.572,-1.363 4.936,0c1.363,1.363 1.363,3.573 0,4.936L66.375,61.439l20.279,20.278c1.363,1.363 1.363,3.573 0,4.937c-1.363,1.362 -3.572,1.362 -4.936,0L61.44,66.376L41.162,86.654c-1.362,1.362 -3.573,1.362 -4.936,0c-1.363,-1.363 -1.363,-3.573 0,-4.937l20.278,-20.278L36.226,41.162c-1.363,-1.363 -1.363,-3.573 0,-4.936c1.363,-1.363 3.573,-1.363 4.936,0L61.44,56.504L81.719,36.226L81.719,36.226z"
+      android:fillType="evenOdd"/>
+</vector>
