+++ app/src/main/res/layout/content_request_list.xml
@@ -1,0 +1,107 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical">
+
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="110dp">
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="45dp"
+            android:layout_marginLeft="45dp"
+            android:orientation="vertical"
+            app:layout_constraintHorizontal_bias="0.0"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintVertical_bias="0.0">
+
+           <LinearLayout
+               android:layout_width="match_parent"
+               android:layout_height="wrap_content"
+               android:orientation="horizontal">
+
+               <TextView
+                   android:id="@+id/sender_text_view"
+                   android:layout_width="wrap_content"
+                   android:layout_height="40dp"
+                   android:paddingStart="5dp"
+                   android:paddingLeft="5dp"
+                   android:paddingTop="5dp"
+                   android:layout_weight="1"
+                   android:textStyle="bold"
+                   android:text="sample sender"
+                   android:textColor="@color/black"
+                   android:textSize="24sp"
+                   tools:ignore="RtlSymmetry" />
+
+               <Button
+                   android:id="@+id/accept_request_button"
+                   android:layout_width="wrap_content"
+                   android:layout_height="wrap_content"
+                   android:layout_marginEnd="10dp"
+                   android:layout_marginRight="10dp"
+                   android:textSize="16sp"
+                   android:clickable="true"
+                   android:focusable="true"
+                   android:text="Accept"
+                   style="@style/Widget.AppCompat.Button.Borderless.Colored"
+                   android:textColor="@color/blue"/>
+
+           </LinearLayout>
+
+            <TextView
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:paddingStart="5dp"
+                android:paddingLeft="5dp"
+                android:textSize="16sp"
+                android:textColor="@color/grey"
+                android:text="sent you a request and said:"/>
+
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:orientation="horizontal">
+
+                <TextView
+                    android:id="@+id/message_request_text_view"
+                    android:layout_width="0dp"
+                    android:layout_height="40dp"
+                    android:layout_weight="1"
+                    android:padding="5dp"
+                    android:text="'Sample messages'"
+                    android:textColor="@color/black"
+                    android:textSize="18sp"
+                    android:textStyle="italic" />
+
+            </LinearLayout>
+
+        </LinearLayout>
+
+        <ImageButton
+            android:id="@+id/dismiss_button"
+            android:layout_width="24dip"
+            android:layout_height="50dip"
+            android:layout_marginStart="10dp"
+            android:layout_marginLeft="10dp"
+            android:layout_marginTop="20dp"
+            android:layout_marginBottom="20dp"
+            android:background="@color/white"
+            android:scaleType="fitCenter"
+            android:clickable="true"
+            android:focusable="true"
+            android:src="@drawable/ic_delete_cross"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent" />
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+
+</LinearLayout>
