+++ app/src/main/res/values/colors.xml
@@ -8,5 +8,6 @@
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
     <color name="blue">#4287f5</color>
+    <color name="grey">#6C6C6C</color>
 
 </resources>
